+++ src/test/java/notai/document/application/DocumentServiceTest.java
@@ -1,5 +1,6 @@
 package notai.document.application;
 
+import notai.pdf.PdfService;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.junit.jupiter.MockitoExtension;
